+++ packages/espressocash_app/lib/features/incoming_split_key_payments/module.dart
@@ -2,13 +2,13 @@ import 'package:flutter/material.dart';
 import 'package:nested/nested.dart';
 import 'package:provider/provider.dart';
 
-import '../../core/accounts/module.dart';
-import '../../core/balances/context_ext.dart';
 import '../../di.dart';
-import 'src/bl/iskp_repository.dart';
-import 'src/bl/tx_created_watcher.dart';
-import 'src/bl/tx_sent_watcher.dart';
-import 'src/widgets/pending_iskp_listener.dart';
+import '../accounts/module.dart';
+import '../balances/widgets/context_ext.dart';
+import 'data/iskp_repository.dart';
+import 'services/pending_iskp_listener.dart';
+import 'services/tx_created_watcher.dart';
+import 'services/tx_sent_watcher.dart';
 
 class ISKPModule extends SingleChildStatelessWidget {
   const ISKPModule({super.key, super.child});