+++ packages/espressocash_app/lib/features/payment_request/widgets/formatted_amount.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/widgets.dart';
+
+import '../../../core/presentation/format_amount.dart';
+import '../../../core/tokens/token_list.dart';
+import '../models/payment_request.dart';
+
+extension PaymentRequestFormatAmountExt on PaymentRequest {
+  String formattedAmount(Locale locale) =>
+      payRequest.cryptoAmount(TokenList())?.format(locale) ?? '';
+}